rename interface and implementation class of KeyStoreAccess
