+++ src/test/java/com/teragrep/jai_02/tests/KeyStoreAccessImplTest.java
@@ -46,30 +46,24 @@
 
 package com.teragrep.jai_02.tests;
 
-import com.teragrep.jai_02.keystore.KeyStoreAccess;
+import com.teragrep.jai_02.keystore.KeyStoreAccessImpl;
 import com.teragrep.jai_02.keystore.KeyStoreFactory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import java.io.IOException;
-import java.security.InvalidKeyException;
-import java.security.KeyStoreException;
-import java.security.UnrecoverableEntryException;
-import java.security.spec.InvalidKeySpecException;
-
-public class KeyStoreAccessTest {
+public class KeyStoreAccessImplTest {
 
     private static String keyStorePath = "target/keystore.p12";
     private static String keyStorePassword = "changeit";
     private static String userName = "trusted-12";
     private static String userPassWord = "XOsAqIhmKUTwWMjWwDaYmVgR8sl_l70H1oDPBw9z2yY";
 
-    private static KeyStoreAccess ksa;
+    private static KeyStoreAccessImpl ksa;
     @BeforeAll
     public static void prepare() {
         Assertions.assertDoesNotThrow(() -> {
-            ksa = new KeyStoreAccess(new KeyStoreFactory(keyStorePath, keyStorePassword.toCharArray()).build(), keyStorePath, keyStorePassword.toCharArray());
+            ksa = new KeyStoreAccessImpl(new KeyStoreFactory(keyStorePath, keyStorePassword.toCharArray()).build(), keyStorePath, keyStorePassword.toCharArray());
             ksa.deleteKey(userName);
         });
     }
